requirements.txt
hydra-core=1.3.2
pandas=2.3.1
tqdm=4.67.1
openai=1.93.2
Pillow=11.3.0
matplotlib=3.10.3
scikit-learn=1.7.0
numpy=2.3.1 